# Optimise display methods