--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,3 +19,5 @@ strum_macros = "0.24"
 chrono = "0.4.19"
 iso8601 = "0.4.1"
 paste = "1.0.7"
+[lib]
+doctest = false