--- a/src/datetime.rs
+++ b/src/datetime.rs
@@ -1,6 +1,5 @@
 use std::cmp::Ordering;
 use std::fmt;
-
 use crate::{get_digit, Date, ParseError, Time};
 
 /// A DateTime
@@ -33,13 +32,66 @@ pub struct DateTime {
 
 impl fmt::Display for DateTime {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}T{}", self.date, self.time)?;
+        //write!(f, "{}T{}", self.date, self.time)?;
+        //this change 129ns/iter => 52ns/iter
+        if self.time.microsecond != 0 {
+            let mut buf: [u8; 26] = *b"0000-00-00T00:00:00.000000";
+            buf[0] = b'0' + (self.date.year / 1000) as u8;
+            buf[1] = b'0' + (self.date.year / 100 % 10) as u8;
+            buf[2] = b'0' + (self.date.year / 10 % 10) as u8;
+            buf[3] = b'0' + (self.date.year % 10) as u8;
+            buf[5] = b'0' + (self.date.month / 10) as u8;
+            buf[6] = b'0' + (self.date.month % 10) as u8;
+            buf[8] = b'0' + (self.date.day / 10) as u8;
+            buf[9] = b'0' + (self.date.day % 10) as u8;
+            buf[11] = b'0' + (self.time.hour / 10) as u8;
+            buf[12] = b'0' + (self.time.hour % 10) as u8;
+            buf[14] = b'0' + (self.time.minute / 10) as u8;
+            buf[15] = b'0' + (self.time.minute % 10) as u8;
+            buf[17] = b'0' + (self.time.second / 10) as u8;
+            buf[18] = b'0' + (self.time.second % 10) as u8;
+            buf[19] = b'.';
+            buf[20] = b'0' + (self.time.microsecond / 100000 % 10) as u8;
+            buf[21] = b'0' + (self.time.microsecond / 10000 % 10) as u8;
+            buf[22] = b'0' + (self.time.microsecond / 1000 % 10) as u8;
+            buf[23] = b'0' + (self.time.microsecond / 100 % 10) as u8;
+            buf[24] = b'0' + (self.time.microsecond / 10 % 10) as u8;
+            buf[25] = b'0' + (self.time.microsecond % 10) as u8;
+            f.write_str(std::str::from_utf8(&buf[..]).unwrap())?;
+        } else {
+            let mut buf: [u8; 19] = *b"0000-00-00T00:00:00";
+            buf[0] = b'0' + (self.date.year / 1000) as u8;
+            buf[1] = b'0' + (self.date.year / 100 % 10) as u8;
+            buf[2] = b'0' + (self.date.year / 10 % 10) as u8;
+            buf[3] = b'0' + (self.date.year % 10) as u8;
+            buf[5] = b'0' + (self.date.month / 10) as u8;
+            buf[6] = b'0' + (self.date.month % 10) as u8;
+            buf[8] = b'0' + (self.date.day / 10) as u8;
+            buf[9] = b'0' + (self.date.day % 10) as u8;
+            buf[11] = b'0' + (self.time.hour / 10) as u8;
+            buf[12] = b'0' + (self.time.hour % 10) as u8;
+            buf[14] = b'0' + (self.time.minute / 10) as u8;
+            buf[15] = b'0' + (self.time.minute % 10) as u8;
+            buf[17] = b'0' + (self.time.second / 10) as u8;
+            buf[18] = b'0' + (self.time.second % 10) as u8;
+            f.write_str(std::str::from_utf8(&buf[..]).unwrap())?;
+        }
         if let Some(offset) = self.offset {
             if offset == 0 {
                 write!(f, "Z")?;
             } else {
+                // let mins = offset / 60;
+                // write!(f, "{:03}:{:02}", mins / 60, (mins % 60).abs())?; //+00:01
+                //this change 105 ns/iter (+/- 1) to 63 ns/iter (+/- 0)
                 let mins = offset / 60;
-                write!(f, "{:+03}:{:02}", mins / 60, (mins % 60).abs())?;
+                let min = mins / 60;
+                let sec = (mins % 60).abs();
+                let mut buf: [u8; 6] = *b"+00:00";
+                buf[1] = b'0' + (min / 10 % 10) as u8;
+                buf[2] = b'0' + (min % 10) as u8;
+                buf[4] = b'0' + (sec / 10 % 10) as u8;
+                buf[5] = b'0' + (sec % 10) as u8;
+                f.write_str(std::str::from_utf8(&buf[..]).unwrap())?;
             }
         }
         Ok(())